[vga_text+tty] vga_init now takes in the width and height in characters. Add vga_scroll function and call it in tty_print.

+++ include/dxgmx/video/tty.h
@@ -10,6 +10,6 @@
 
 void tty_init();
 void tty_clear();
-int tty_print(const char *str, size_t n);
+size_t tty_print(const char *str, size_t n);
 
 #endif // _DXGMX_VIDEO_TTY_H
